added Entertainment Repository
