+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/EntertainmentViewModel.java
@@ -4,6 +4,7 @@ import android.app.Application;
 import android.util.Log;
 
 import com.sdzshn3.android.news247.Fragments.EntertainmentNewsFragment;
+import com.sdzshn3.android.news247.Repositories.EntertainmentRepository;
 import com.sdzshn3.android.news247.Retrofit.ApiService;
 import com.sdzshn3.android.news247.Retrofit.Article;
 import com.sdzshn3.android.news247.Retrofit.Client;
@@ -20,54 +21,19 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 public class EntertainmentViewModel extends AndroidViewModel {
-    private static MutableLiveData<List<Article>> data = new MutableLiveData<>();
 
-    private final ApiService apiService;
-    private static Call<NewsModel> call;
+    private EntertainmentRepository repository;
 
     public EntertainmentViewModel(@NonNull Application application) {
         super(application);
-        apiService = Client.getApiService();
-        loadData();
+        repository = new EntertainmentRepository();
     }
 
     public LiveData<List<Article>> getData() {
-        return data;
-    }
-
-    private void loadData() {
-        call = apiService.getResponse(EntertainmentNewsFragment.URL);
-        call.enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("EntertainmentViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        return repository.getData();
     }
 
     public void refresh() {
-        call = apiService.getResponse(EntertainmentNewsFragment.URL);
-        call.clone().enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("EntertainmentViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository.refresh();
     }
 }
